backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg[binary]==3.2.2
python-decouple==3.8
django-cors-headers==4.4.0
django-filter==24.3
gunicorn==22.0.0
whitenoise==6.7.0
drf-spectacular==0.27.2
